+++ releasenotes.md
@@ -1,5 +1,10 @@
 # 0.4.8
 
+- Don't show checkbox for failed challenges
+- Sort all challenges by latest at date
+
+# 0.4.8
+
 - Add sorting by failing date for overdue challenges
 
 # 0.4.7